fix: style

+++ src/Utils/Log.cpp
@@ -41,14 +41,10 @@ static struct {
     Callback callbacks[MAX_CALLBACKS];
 } L;
 
-static const char *level_strings[] = {
-    "TRACE", "DEBUG", "INFO", "WARN", "ERROR", "FATAL"
-};
+static const char *level_strings[] = {"TRACE", "DEBUG", "INFO", "WARN", "ERROR", "FATAL"};
 
 #ifdef LOG_USE_COLOR
-static const char *level_colors[] = {
-    "\x1b[94m", "\x1b[36m", "\x1b[32m", "\x1b[33m", "\x1b[31m", "\x1b[35m"
-};
+static const char *level_colors[] = {"\x1b[94m", "\x1b[36m", "\x1b[32m", "\x1b[33m", "\x1b[31m", "\x1b[35m"};
 #endif
 
 static void stdout_callback(log_Event *ev) {
@@ -56,14 +52,11 @@ static void stdout_callback(log_Event *ev) {
     const auto elapsed_ms = std::chrono::duration_cast<std::chrono::milliseconds>(now - start_time).count();
 
 #ifdef LOG_USE_COLOR
-    fprintf(
-        static_cast<FILE *>(ev->udata), "[%5ldms] %s%-5s\x1b[0m \x1b[90m%s:%d:\x1b[0m ",
-        elapsed_ms, level_colors[ev->level], level_strings[ev->level],
-        ev->file, ev->line);
+    fprintf(static_cast<FILE *>(ev->udata), "[%5ldms] %s%-5s\x1b[0m \x1b[90m%s:%d:\x1b[0m ",
+            static_cast<long>(elapsed_ms), level_colors[ev->level], level_strings[ev->level], ev->file, ev->line);
 #else
-    fprintf(
-        static_cast<FILE *>(ev->udata), "[%5ldms] %-5s %s:%d: ",
-        elapsed_ms, level_strings[ev->level], ev->file, ev->line);
+    fprintf(static_cast<FILE *>(ev->udata), "[%5ldms] %-5s %s:%d: ", static_cast<long>(elapsed_ms),
+            level_strings[ev->level], ev->file, ev->line);
 #endif
     vfprintf(static_cast<FILE *>(ev->udata), ev->fmt, ev->ap);
     fprintf(static_cast<FILE *>(ev->udata), "\n");
@@ -82,27 +75,21 @@ static void unlock() {
     }
 }
 
-const char *log_level_string(const int level) {
-    return level_strings[level];
-}
+const char *log_level_string(const int level) { return level_strings[level]; }
 
 void log_set_lock(const log_LockFn fn, void *udata) {
     L.lock = fn;
     L.udata = udata;
 }
 
-void log_set_level(const int level) {
-    L.level = level;
-}
+void log_set_level(const int level) { L.level = level; }
 
-void log_set_quiet(const bool enable) {
-    L.quiet = enable;
-}
+void log_set_quiet(const bool enable) { L.quiet = enable; }
 
 int log_add_callback(const log_LogFn fn, void *udata, const int level) {
     for (auto &callback: L.callbacks) {
         if (!callback.fn) {
-            callback = (Callback){fn, udata, level};
+            callback = (Callback) {fn, udata, level};
             return 0;
         }
     }
@@ -119,10 +106,10 @@ static void init_event(log_Event *ev, void *udata) {
 
 void log_log(const int level, const char *file, const int line, const char *fmt, ...) {
     log_Event ev = {
-        .fmt = fmt,
-        .file = file,
-        .line = line,
-        .level = level,
+            .fmt = fmt,
+            .file = file,
+            .line = line,
+            .level = level,
     };
     lock();
     if (!L.quiet && level >= L.level) {
